+++ django_nanopages/__init__.py
@@ -1,4 +1,5 @@
 from .pages import Pages  # noqa
-from .views import Page  # noqa
+from .page import Page  # noqa
+from .views import PageView  # noqa
 
 __version__ = "0.2.0"